refactor: Calendar 타입 이동
1. Calendar 타입을 types 폴더로 이동
#17
